# apps/api/src/main/java/com/example/authapi/dto/request/TokenRefreshRequest.java
package com.example.authapi.dto.request;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class TokenRefreshRequest {
    
    @NotBlank(message = "Refresh token is required")
    private String refreshToken;
}